AbstractMarker
Added ‘AbstractMarker’ class and a vector of abstractMarker to the
tgModel class.

Abstract Markers are the non-physical entities that are attached to a
specific rigid body from a specific point. They provide exact location
of that point while the rigid body moves and rotates.

They are visualized by spheres in the graphics simulation. They are
also used to receive positions for control algorithms.