SuperBallModel reconstructred and moveModel added
SuperBallModel is reconstructed so that it will have different
numbering for nodes. I also added helper functions and arrays to find
corresponding rods or nodes (such as what is the node that is at the
end of the same rod?)

Added moveModel function so that after the model is created, user can
manually give position, velocity and rotation to the structure. This is
necessary for controls experiments.