Minor refactoring of SpineHardware/VerticalSpine model file in preparation for switching over to pretension from setRestLengthSingleStep. This commit should be squashed by Drew later.
Second code refactor for VerticalSpine.
This commit also updates the VerticalSpine model to use pretension instead of setRestLengthSingleStep.
This commit should put VerticalSpine up-to-date with the rest of the repository for release v1.1

This squash combines all Drew's changes to dev/SpineHardwareProject to remove depreciated code and refactor the spine's model class.